TD1: 2.2: Building our application
For the moment, the build isn't doing much apart from displaying two messages. So we're going to modify the pipeline to build our java application.

To do this, modify the .github/workflows/main.yml file to add the following build steps:

      # Installe java 8 sur l'agent
      - name: Set up JDK 8
        uses: actions/setup-java@v2
        with:
          java-version: '8'
          distribution: 'adopt'

      # Compile notre application
      - name: Build with Maven
        run: mvn --batch-mode --update-snapshots package
Start your changes and observe the build of your application.