2.3: Integrate test results
If you look closely at the build logs, you will see that unit tests are being run. We are going to retrieve the test report and display it in the pipeline result.

To do this, modify the .github/workflows/main.yml file to add the following build steps:

      # Publie les résultats de test
      - name: Publish Unit Test Results
        uses: EnricoMi/publish-unit-test-result-action@v1
        if: always()
        with:
          files: target/**/*.xml
Start your changes and observe the build and the results of your application's unit tests.