2.4: Publishing the built artifact
You are now going to publish the artifact you built in the previous steps. This allows you to download the artifact in question and also to deploy it in later stages on different environments.

To do this, modify the .github/workflows/main.yml file to add the following build steps:

      # Publie les fichiers présents dans le dossier target dans un build artifact
      - name: Upload a Build Artifact
        uses: actions/upload-artifact@v2.2.4
        with:
          name: my-app-1.0
          path: target/*
Commit your changes and observe the build and the artifact published on the build page.